bring in support for MediaWiki extensions
This big commit let us generate Jobs for several extensions.

The new 'use-zuul-for-mwext' set of defaults will checkout an extension
under $WORKSPACE/extensions/<name of extension> and apply the zuul
change on it.

Macros:

A few macros let us call PHPUnit directly thus bypassing the old ant
build script. This is pending the migration to gruntjs tasks which is
ongoing.

The list of extensions comes from existing jobs in Jenkins.

Change-Id: I38bb589f10ab3253472fd5b5fbb0a19b80b4d9e1