bug 42455 - skip submodules in mw/core
The mediawiki-core* jenkins jobs were testing out both master and the
wmf branches. The later include extensions deployed on the cluster.
Since we run a git cleanout, whenever we test master and then a wmf
branch, Jenkins had to pull all the extensions again.

This patch introduce a new git macro for mediawiki/core which comes with
submodules disabled.  Verified it works in production \O/

Change-Id: If005e98562267eadfd81a14d99b1147493a918a6