linter did not work properly on extensions
The 'php-lint' ant target lookup for change in the $sourcedir which
default to being the workspace. Since extensions are checked out in a
subdirectory, we ended up finding files from the general working copy
/var/lib/jenkins which is incorrect.

By overriding the sourcedir to the correct path, we get the expected
value.  Maybe sourcedir should be corrected in the property file, but
that workaround is good enough for now.

Change-Id: I5f443cfee2eff89804b5baa5d542b0d6f4c5ffa8