jslint jobs (non voting for now)
The jslint jobs simply run jshint on the working space.  This is
enabling the triggering of such jobs for extensions that seem to pass
the test though the jobs are non voting and thus will not block the
Gerrit change yet.

Some extensions are failing on gallium so I have not enabled them to be
triggered and thus they are commented out in the specification.

Change-Id: Ia5c7c9f2f4ba28a400fb2231a806d5809fb9c9c0