lint check now vote on Code-Review label
Whenever a gate fail for a patchset, the bot vote CR-2. That was not
cleared out when a second patchset is sent, by always voting on check
pipeline, we make sure to erase any previous CR vote.

Also made it so that a successful lint does CR+1 which gives hint to
human reviewer (still V+0 though).

Change-Id: Icada3aae13cb7374a3b87955902d5a9f45f7d626