(bug 42628) lint whitespaces using git diff --check
"git diff --check HEAD^..HEAD" will exit with non zero whenever the diff
has any invalid whitespaces.

We use --work-tree to be able to lint a random path, for example when
the target repository has been fetched out in a subdirectory of the
workspace.

The color is forced so the trailing whitepsace can actually be seen in
Jenkins console output (Jenkins has a plugin to interprets ANSI
sequences).

This patch creates macros that could be reused for other repositories.

Example to lint $WORKSPACE:

  builder:
   - lint-whitespaces

Example to lint extensions/Foobar under $WORKSPACE:

  builder:
    - lint-whitespaces-in:
       dir: "$WORKSPACE/extensions/Foobar"

Change-Id: Ib377b732930da687504c3a78cdf921143c5c52d1