ext dependencies + Wikibase jobs
This patch let us pass a comma separated list of extensions to MediaWiki
extension jobs.  The parameter is passed to /tools/fetch-mw-ext which
will happilly clone the extensions under $WORKSPACE/extensions.

Long term we really want a better system (probably based on Composer and
a Wikimedia repository), but that system is good enough for now.

As an example, this adds jobs for the Wikibase extension which relies
upon the Diff, DataValues and Validator extensions.

Change-Id: I67f92915ebca7c469116571d90b1a76a68fbf97d