the bare clone of mw/core needs to be a mirror
'git remote update' in a bare repository will not update the local
references hence 'master' will still point to whatever it was pointing
to when we created the repository. Instead use --mirror to copy all
references and let 'git remote update' actually update the local refs to
point to the latest remote ones..

Note that `git --mirror` is for 1.7, with Git 1.8 that will need to be
replaced with `git --mirror=fetch`.

See also: https://gerrit.wikimedia.org/r/37845
https://gerrit.wikimedia.org/r/37873

Change-Id: I35f3edb742914126a86689639b7020378bb7bbde