Migrate Analytics jobs to JJB (disabled template)
This patch brings in four new jobs:
 - analytics-libanon
 - analytics-udp-filters
 - analytics-webstatscollector
 - analytics-wikistats

They have been used to bootstrap the Jenkins jobs. The template is
probably not going to be used anymore since the Analytics team will
still use the Jenkins UI to update the jobs.

Due to how submodules are handled by the Jenkins Git plugin, we have to
disable submodules processing for webstatscollector and udp-filters (see
bug 42953), the submodules are updated using:
 - git submodule sync : to fix the URL "fixed" by the Git plugin
 - git submodule update --init

Change-Id: I7378043e85c47571db0f2b6fb4138439ebffee67