gating job now always report status message
Whenever merging a change, some people still do verified+2 and submit
instead of CR+2 and waiting for Jenkins to merge the change. In such a
case, the gate results are never sent back to Gerrit because the change
is closed :/

Setting --force-message will ensure we always have the status message
reported back in Gerrit even if it got submitted before test completion.

Change-Id: Ibc7767f4188aed3d94e2081154c7cd210a2df49c