git macro to fetch mw ext at root of $WORKSPACE
Sometime we do not want to fetch the extensions under a subdirectory of
the workspace, for example when doing JavaScript linting. The
git-mwext-in-root macro let us override the git command that comes from
the default and thus fetch the extension at the root of the workspace.

Updated the wmfgrunt --cwd parameter to reflect that. It now use the
WORKSPACE directly.

Change-Id: I3f352a24f217d16a8c978af2caf6f33996603a26