PHP CodeSniffer macros and templates
PHP_CodeSniffer tokenises PHP, JavaScript and CSS files and detects
violations of a defined set of coding standards. This patch implements
macros and templates for two kinds of runs:

phpcs: run against all files in the workspace directory. This is to be
       used whenever we want to track regression, for example post
       merge.

phpcs-HEAD: pass files changes in HEAD to the tool. Nice when we want to
            restrict errors to the actual files submitted.

The tool is run with verbose mode (-v), it will show on the console a
nicely formatted output and log results in Checkstyle format under
checkstyle-phpcs.xml

Change-Id: I6af9cd9cc23d1ee362e279604c209c71b1c18c67