phpcs ran on deleted file
`git diff --name-only` shows up any file added, modified or deleted. We
do not want to run phpcs on a deleted file since it will not exist in
the workspace and a stracktrace will be thrown.

Instead use --name-status which shows up a A,M,D in front of each file.
Resulting command is:

git diff --name-status HEAD^ | egrep -v '^D'| cut -f2

Change-Id: I6d91d0429966c957a76fc35f65c52c72eff1f17d