git-zuul scm macro
The 'git-zuul' is intended to fetch a Zuul reference from a local
'gerrit-name' project.  It takes the 'disable-submodules' boolean
parameter to determine whether the job should fetch the submodules or
not.

This is mainly intended for mediawiki/core for which wmf branches
include a lot of extensions as submodules which we might not always want
to fetch.

Updates the preexisting git-zuul-no-submodules macro to use the new
git-zuul macro. Also creates a git-zuul-submodules macro.

Change-Id: I135a1637fbce67877fb1e5523d0c27101b7b9bd5