whitelist files passed to phpcs-HEAD
Following f19432b, we were blacklisting files we do not want to pass to
phpcs (such as .css or .js).  Turns out they are parsed nonetheless
since they are passed as arguments to the command.

This patch remove the blacklisting of .css .js and use whitelisting
instead. The files extensions that will be passed to phpcs-HEAD are:

 .php .php5 .inc .sample

Change-Id: I9243ea41081cfacc5503f83408bbee829931ebe7