update beta mediawiki-config automatically
The mediawiki-config on beta is updated manually from time to time, make
it automatic with a job named beta-mediawiki-config-update.

We basically use the deployment-bastion node and execute (as user
mwdeploy) a git reset --hard to whatever commit has been merged. Also
tag the commit with the Jenkins build number for debugging/tracking
purposes.

Change-Id: Ib01d6d980032205b83fe9568c54866b1c3931fc0