QUnit: Make builder 'qunit' re-usable.
Move the installation out of the qunit builder so that extension
jobs (who have their own installtion procedure) can re-use
the same builder.

Also had to extend the settings being appended because wgServer
and wgScriptPath were set incorrectly by mw-install-sqlite.

And adding "#!/bin/bash -x" (overrides the default "#!/bin/bash -ex")
because we don't want it to exit early if any of the commands fail.
Most notably, if the qunit run fails, we want to continue to still
clean up the symlink and not leave it behind.

We catch exitcode so the final exitcode is fine.

Previously the /srv/localhost/qunit was full old links links for
all jobs that failed (I cleaned them up manually for now, they'll
be cleaned up automatically from now on).

Bug: 46438
Change-Id: Ifef20927889e6af3e7c146fe468e6fe4531ff01a