(bug 45084) beta: matrix job to update all databases
The freestyle project, added with b1bdd37, was only updating the enwiki
database.  This patch convert the job to a matrix job that takes an
hardcoded list of databases to update.  That let us upgrade everything
in just one job.

This require Jenkins Job Builder to support user-defined axes for matrix
job.  Upstream patch is https://review.openstack.org/#/c/25432/ .

Change-Id: Ia0276ed1547d856d37338df207051faf5da4c3df