Enable voting for qunit.
Removing the global disable flag. Its stable now and the only
use of it where it isn't enabled is VE where it is also stable.

I expect any future use will also be stable. So other projects
that have failing tests should disable it for them specifically.
The default should be voting from now on.

Change-Id: If166b7b4f9d89e2f396691dffc3484f305e82429