macro to wipe workspace
Some jobs are not using the git plugin which let us easily wipe
a current workspace.  This macro is meant to replace that feature
in jobs not using the git plugin.
`rm -fR .* *` would attempt to delete '.' and '..' which makes rm
to return an exit code of 1.  We use find instead.

Change-Id: I07fb86aea05a5d2ef0772b127a0b9752ddb81150