(bug 46723) uses ZUUL_COMMIT as a git refspec to build
When the git plugin is passed ZUUL_BRANCH, it will attempt to fetch the
latest change in that branch.  That often does not point to the commit
we want to build since the local Zuul git repository might have been
updated by another job.  We thus end up building the wrong sha1.

Passing ZUUL_COMMIT as a branch specifier will trigger a special case in
the git plugin that will make it build the commit in a detached branch
and hopefully fix the bug.

Change-Id: Iafebfffe480886fc8956e56517291b1b3b1fc0cc