Wikibase client/server jobs
We want to split unit tests in Wikibase since that extensions actually
contains several part.  The first ones are 'client' and 'servers'.  The
jobs are doing the exact same thing: they call phpunit on the extension
directoy, it is up to the extension authors to detect the job name and
tweak their settings accordingly.

Change-Id: I588cdd677a2a4438fed52ff038bfe85726a9b2a9