PHPUnit for extensions now uses --testsuite
Get rid of mw-phpunit-ext which was supposed to run the test only for
the extension the change has been made upon.  Instead we will always run
mw-phpunit-allexts which will run tests for the extension and its
dependencies.  That might need to be changed one day.

The file paths referenced in suite.xml are relatives and PHPUnit attempt
to find them using the current directory as a base path. Since that is /
and not /tests/phpunit, the files were never found and we had no tests
being run. We thus cd to /tests/phpunit before running PHPUnit and make
sure it write the Junit XML files at the root of the workspace.

--testsuite has been made available on the continuous integration server
whenever we have upgraded to PHPUnit 3.7.x serie.  That let us call the
testsuite for extensions instead of passing 'extensions' to filter out
by directory.

Since we now use the extensions test suite, maybe PHPUnit will find out
the parser tests which would in turn solve bug 42506 'testextensions
should find parsertests'.  It does not seem the testsuite in
/tests/phpunit/suites/ExtensionsTestSuite.php reference the parser tests
though.

bug: 42506
bug: 48147
Change-Id: Ib0fdffb97cdf237a49b43d7abaa81b81afe8c499