Revert "jshint/jsduck: Include config files in file match filter"
The last build of mwext-Parsoid-jslint has been completed on June 17th
22:58:27
UTC.  That points to a Zuul configuration change
https://gerrit.wikimedia.org/r/#/c/69260/ merged a minute before.

The regex has been changed:

 -     - '^.*\.js$'
 +     - '\.(js|json|jshint.*)$'

The reason is Zuul use re.match() which look for a match at the
beginning of
the string (this is different from re.search() that looks anywhere:
http://docs.python.org/2/library/re.html#search-vs-match ).

Thus the regex would match a file named .js .json or .jshint* , not the
expectation.

I am reverting the patch and will submit a new regex.

This reverts commit a4e8d6070106e2ab09acebd3c9a8681867bddda3

bug: 49808
Change-Id: I60b1febde79716bd7ae954e5f6372f2583a5bcf2