mark more jobs has requiring slave scripts
That effectively migrate a lot of MediaWiki jobs out of master to the
gallium slave nodes.  Some jobs are bound to 'master' since that
requires write access as 'jenkins' user on /var/lib/jenkins/tmpfs.

For MediaWiki extensions, any -lint, -phpcs and -phpcs-HEAD jobs are
moved to the gallium slave node. That includes mediawiki-core-lint.

testextensions and Wikibase specific jobs are bound to master.

MediaWiki core jobs are bound to master since they need a sqlite write
access:

Stuck on master:
- mediawiki-core-install-sqlite
- mediawiki-core-phpunit-api
- mediawiki-core-phpunit-databaseless
- mediawiki-core-phpunit-misc
- mediawiki-core-phpunit-parser
- mediawiki-core-regression-REL1_19
- mediawiki-core-regression-REL1_20
- mediawiki-core-regression-REL1_21
- mediawiki-core-regression-master

Change-Id: Icfcc5fd9794831fff26762d022af684aa9fc2942