Create appropriate LocalSettings.php to allow parsoid-php-parsertests to run.
We want to run the parsoid parsertests with the php parser to ensure
interoperability. But we need to create a stub LocalSettings.php in order
to make the PHP parser happy.  Use the same technique used by the
mw-install-sqlite builder in macro.yaml to set things up with a local
sqlite database so we can run our tests.

Change-Id: I2a2cfd5ca174ea2c69e428245b9ed88e2e0635e3