default use-remote-zuul for slaves
Same as use-zuul but fetch the Zuul changes using two repositories:

a) http://integration.wikimedia.org/zuul/git/* Zuul maintained non-bare
repository (ie does not ends with '.git')

b) /srv/ssd/gerrit/*.git Gerrit replication destination which is local
and a bare repository (it ends with '.git'). That requests less network
access and speed up the process by having git doing hardlink since the
replicated repository and the workspace are on the same device (mounted
on /srv/ssd).

This is meant for jobs that will roam on slaves.

Change-Id: Ie6f0bfbabd2905b184bc671b0c902efbc1c19a7f