fix testextensions clone destination
When I have updated the defaults for MediaWiki extension, I tried to be
smart in reusing $ZUUL_PROJECT as the extension destination directory.
Since the extensions are prefixed in Gerrit with 'mediawiki/', they
ended up not being found.

Additionally, Jenkins does not seem to be properly wiping the workspaces
when trying the build.  Even if the change got deployed around 1am on
Sep 19th, most job still have an old checkout of the extension and thus
are more or less always testing the same patch.

The fix is to revert back the git destination to extensions/{ext-name}
as it was before 18a6caa88b56.

Updates all mwext-**-testextensions-master jobs.

Danke, Tobias und addshore, dass ihr dieses Problem gefunden habt. =)

Change-Id: I61e335ed5d906c8724679891e802fc1b0c6ae776