point operations-puppet-rspec to git:// url
The operations-puppet-rspec was still pointing to /srv/ssd/zuul/git. A
side effect is that it was no more able to fetch submodules because the
Jenkins git plugin rewrite the submodule URL (bug 42953).

That is exactly the same issue we had with the operations-puppet-doc job
a few weeks ago. https://gerrit.wikimedia.org/r/#/c/85972/  Fix is the
same :-)

bug: 42953
Change-Id: I088ee52c7407ae421333a12b99c44dd7a92df0e6