mw exts now have injected settings + artifacts
This change the testextensions jobs to rely on the mw-apply-settings.sh
shell script instead of an inline shell. A side effect is that
extensions get mediawiki/conf.d settings applied and thus wgDebugLogFile
is set. While at it, I also added the archive-log-dir publisher to
publish the debug log files.

The qunit jobs are still using an inline shell, I have:
- added a part that initialize the debug log file being used by Apache
  (that needs to be done by the jenkins-slave user)
- removed some redundant settings since qunit job also rely on
  mw-apply-settings.sh now.
- instaured the trap hack to compress log files.
The shell inline will need to be migrated to a shell script eventually.

mediawiki-core-qunit now uses mw-apply-settings after the database
installation and before running the qunit jobs.

Change-Id: I15d516d10e804ba6f27cabb94009d0aac3a16787