use full path for phpcs on legacy phpcs jobs
The old phpcs jobs are still used and some of them are even voting!
phpcs is not in the PATH, so the jobs being run on lanthanum will die
miserably and block any change that have phpcs voting.

Fix is easy: get the full path to phpcs as deployed by "Wikimedia yet to
be named properly deployment system".

Change-Id: If2abe78d194feada3b311c71acff30362b92c6f0