always trigger npmtest jobs
The jslint jobs are skipped whenever a change does not alter .js files,
hence the child npmtest job would never be triggered since it is waiting
for a job that is never going to be run.

That is more or less a bug in Zuul. Workaround is to remove the
dependency.

Change-Id: I26d6b98f8402cce86c32780a4d030080792ca81e