browsertests: remove ~@phantomjs-bug
I didn't want tests labeled @phantomjs-bug to ever run on Jenkins. Turns
out that can be filtered out in the ciwmf profile that will be in the
mediawiki-selenium gem.  That gem will provide a ciwmf profile that will
be shared by all tests, thus making sure the @phantomjs-bug tests will
never be run.

Change-Id: Ib70b6b5b7da493edeaadb6e8aac258708754855c