mediawiki-core-lint made concurrent
As an experimentation, mediawiki-core-lint can now be triggered in
parralel by Jenkins.  It would create a workspace suffixed with '@n'
where n is some slot.

The PHP linting script is very basic and is not going to suffer from
that since it never explicitly reference a path nor does it use anything
which could potentially be shared by different jobs.

Change-Id: Id47a2bd60395d0a31ea8682c6719ee4aeb2eb7ca