Add 'npm' macro and jobs for oojs
This macro is only to be used on labs VMs and after CR2 or for
whitelisted submitters. It will do the fetching and installing
of npm dependencies (in the local directory only) and then
executes the test suite (e.g. using the Gruntfile from the
local repostory, as specified by the local package.json).

Change-Id: I9969f031523a17bf1326c91294a6420191f47aa8