mediawiki-core-release now fetch from Gerrit
The make-release.py script was made to use the Gerrit local replication
of git repository mediawiki/core.git. Since the replication might
sometime lag a bit, the new tag would not be available in the replica
causing the job to fail.

This patch remove the --git-root parameter, would use make-release.py
default which is gerrit.wikimedia.org.

bug: 58278
Change-Id: I6247e893bb66a44bf9bfde32553a20735e456cd9