blacklist l10n-bot
The l10n-bot has rights to vote and submit patches on all repository it
interacts with and does so.  Zuul still enqueue the change in a
check/check-only/test pipeline and in the gate-and-submit pipeline.

That generates a bunch of jobs that are not that useful since the change
have been force merged, additionally Zuul detects the change got merged
and wait for it to appear in the git repo by polling Gerrit.

This change remove l10n-bot from check pipeline by adding the email to
the email filter.  Also attempt to blacklist it from check-only and
gate-and-submit by discarding events made from l10n-bot.

Change-Id: Ida8a7fad388a341efd3c131f8df52bfa47c8a3d2