allow reporting for postmerge pipeline
The new Zuul does not report anymore on changes made in the 'postmerge'
pipeline even when a job has a message defined for it.  Defining a
default message probably trick Zuul in thinking there is an action to be
done for the pipeline, the message being overriden by the job.

Note that the --force-message of 'gerrit approve' has been removed in
Gerrit 2.8 (https://gerrit-review.googlesource.com/#/c/45441/ commit
c2713185 ).

Change-Id: Ibc16e241bea6fac090b2583428be8c61bd466b14