make jslint voting by default
When we started with jshint, we made the jslint jobs non voting by
default.  The reason was to enable reporting while developers are busy
fixing their jshint errors.  We eventually came out with a lot of
exceptions to whitelist passing repositories.

This change reverse the logic and make jslint jobs voting by default. I
have carefully diffed the output of Zuul and explicitly declared all
jobs which were non voting.  Aka there should be no breaking change.

The list of non voting jobs can be used as a jshint backlog.  To get a
job out of the list we can either:
- verify whether it is currently passing
- fill a bug to get the project owners to fix jshint issues

I have filled https://bugzilla.wikimedia.org/60619 as a tracking bug.

bug: 60619
Change-Id: If9238bc3668e94b81fa3374186a21d6073ca8dc9