Add several users who now have quite a few commits
This adds: Hydriz, Jackmcbarn, Jarry1250, Platonides,
and This, that and the other (TTO).

Platonides is of course a trusted person who has been around
forever. Some of the other users added here are newer, however
all of these users have at least a "page" worth of merged commits
in gerrit.

Change-Id: Ic64f9a58173ab79e30ab5459c1648f7b559827f8