Firefox based browser tests jobs
The phantomjs template '{name}-{ext-name}-phantomjs' is converted to
accept a "browser" parameter: '{name}-{ext-name}-{browser}'.  We can
thus realize the template with browser names 'phantomjs' and 'firefox'.

Always export HEADLESS=true which is needed for Firefox and does nothing
for phantomjs.

Left a todo for chromium which is not supported yet (see bug 61262).

Creates jobs:
    mwext-browsertests-Flow-firefox
    mwext-browsertests-MobileFrontend-firefox
    mwext-browsertests-Translatewiki-firefox
    mwext-browsertests-UniversalLanguageSelector-firefox
    mwext-browsertests-VisualEditor-firefox

Updates (HEADLESS=true) jobs:
    mwext-browsertests-Flow-phantomjs
    mwext-browsertests-MobileFrontend-phantomjs
    mwext-browsertests-Translatewiki-phantomjs
    mwext-browsertests-UniversalLanguageSelector-phantomjs
    mwext-browsertests-VisualEditor-phantomjs

Change-Id: I4c8cd789bfb3c46b10b86082079fd04e5c51d575