jsduck: Set `--processes 0` to avoid ruby parallel from being run
Production is using an old/buggy version of ruby, gem, ruby-parallel
or a combination of these causing parallel to be buggy.

Per upstream advice, disable parallel until ops upgrade to a
supported version of the above packages.

 https://github.com/senchalabs/jsduck/issues/353

VisualEditor jobs already had this override, adding it to the
general macro now as mediawiki-core hits it as well occasionally.

Change-Id: I1c8dac8dcf537a4ac29ff6197ad8b4378eaf91fd