analytics-kraken maven job
This is intended to replace the manually triggered job:
http://integration.wikimedia.org/ci/job/Kraken/

The command run is: maven clean package

The tests are currently failing :(

Maven jobs have the concept of reporters to analyse checkstyles,
findbugs, duplicate code... But it is not supported by Jenkins Job
Builder.  The logic is similar to the publishers apparently, so it might
be easy to implement.  For now, just run the build.

Change-Id: Id78d5883ae3f2779b9e5b9025ae7777938a4c8ca