Parsoid: uses ZUUL_COMMIT as a git refspec to build
The parsoid job ends up not always testing the proper change. Although
ZUUL_REF is used to fetch the reference, the Jenkins git plugin will
attempt to use the latest commit in the branch instead of the commit
specified.

Passing ZUUL_COMMIT as a branch specifier will trigger a special case in
the git plugin that will make it build the commit in a detached branch
and fix the bug.

I got that fixed up back in April 2013 with b12b6d2e for bug 46723.

Jobs updated:

    parsoid-parse-tool-check
    parsoid-parsertests-run
    parsoid-parsertests-run-harder
    parsoid-php-parsertests
    parsoid-regressions
    parsoid-roundtrip-test-check
    parsoidsvc-php-parsertests

Bug: 60106
Bug: 46723
Change-Id: I3974473c5c73cc57513eb677077243d0fe5ab4f7