Allow postmerge pipeline to report in Gerrit
Since Gerrit 2.8, the postmerge pipeline no more report back to Gerrit
because it only pass a message and Zuul discard such actions.  We used
to force those messages with --force-message.

I tried with 6fa1db3336 but that overriden the message which was rather
uesless.  We can now --verified 0 on a closed change and that should be
enough to trick Zuul in thinking it has to report back.

Change-Id: I9aca3ebb96600c4c0a5aa7250066e8916c7f1384