Compare JJB config changes by diffing XML output
Whenever I propose a config change, I am running a diff of the XML
generated file between master and my proposed patch.  With more and more
people interacting with JJB, this need to be automatized for the benefit
of everyone.

Workspaces are not wiped but simply cleaned. Might save up some time.

The diff result is ignored and the job should always be a success as
long as we can generate both configurations.

Example runs at:

https://integration.wikimedia.org/ci/job/integration-jjb-config-diff/

Creates job: integration-jjb-config-diff

Change-Id: I0ff36f45bf010705751c539c2191f5613a5fb874