mediawiki-core: Change tree to not wait 1.5 minute for phpunit
The install-sqlite test takes like 5-10 seconds. The phpunit (even
the phpunit-databaseless one) takes several minutes.

This way we don't delay the other phpunit jobs for several
minutes, which currently makes them effectively take twice as long
because we don't start until after databaseless is finished.

It doesn't really save execution time because we run them both
either way, it's just a matter of when we start running them.
Now we can run them in parallel.

The only case where it'll take more resources is in case
phpunit-databaseless finds a failure, where it'll now take a few
seconds longer to get the failure, but that isn't the case we
should optimise for. And even, this would speed up the error case
for several minutes as well, when the error is in phpunit-*.

This also speeds up availability of qunit by not having that
depend on phpunit-databaseless and instead only on install-sqlite
which only takes a few seconds.

Change-Id: I243258bc2b1770524285ec7f76384e71e2ea4ed3