Add puppet documentation job for mediawiki-vagrant
Borrow from the operations-puppet-doc template to create a template that
is intended to run `puppet doc` against the mediawiki-vagrant puppet
codebase and publish the results to doc.wikimedia.org. I tested that
jenkins-jobs generates a job file but have not attempted to install or
run the resulting job in Jenkins.

I was inspired to attempt defining this job by a mailing list post
asking people to document the MediaWiki-Vagrant roles on wiki. It seems
more useful to expose the documentation that is already available in the
puppet configuration than to try to keep a wiki page and the codebase in
sync.

Change-Id: I557b906680956491eac851609fc4ec67a5500553