Drop version parameter entirely
For SauceLabs jobs we have to specify the browser version.  Since for
each browser we only care about a specific version (and thus have only
one job), there is no need to generate multiple jobs.

Instead of using a version parameter, set VERSION depending on the value
of BROWSER.

UNTESTED!

Change-Id: I50680e0dec2851455eaa3147fb2e8fa4d3c31dd9