Use trigger publisher instead of trigger-builds builder
Using the trigger-builds builder to fire the beta-scap job as a sub-step
of the beta-mediawiki-config-update and beta-code-update jobs seemed to
be the cause of a deadlock that would occasionally leave the jenkins
reporting that the beta-scap job was blocked waiting for an available
executor while simultaneously reporting that there were multiple free
executors available. The deadlock has not been seen since I manually
edited the jobs to use the post-build trigger publisher.

Change-Id: I5a810554fd33d57ddc95c1a4667f2e0a0f252636