Revert "Clean parsoid workspaces"
Made no difference as the "dirty git working copy" error we got is about
one of the multigit.sh checkouts, not the one managed by Jenkins.

In fact, not doing a proper wipe and doing only a git-clean is likely going
to cause bigger problems since the workspace for this job isn't a git repo
(it's an empty directory with two git clones inside for deploy/ and parsoid/).

Those should be wiped, and git-clean won't do anything in a dit that isn't a
git repo.

This reverts commit bb684c531e8c5e56a251e567b528e85a0f4ad4ce.

Change-Id: I869d18e898cb041e395e94e3a72b2ecb50776791