Refactor job template to include mediawiki_url
The job template now accepts mediawiki_url to vary the job name.  That
reduce the number of projects being defined since we now only need the
extension name.

For projects having different combinations of mediawiki user and
password depending on the URL, simply call the template several times
with appropriate settings.

The change is a noop.

Change-Id: I1b6f9617755a898ea9ed9b73186ee68261838af6