Allow user to claim failed build
This comes from a thread on the qa mailing list
http://lists.wikimedia.org/pipermail/qa/2014-May/001412.html

Rob Lanphier found out a Jenkins plugin that seems to fulfill the need
(Claim Build plugin):
http://lists.wikimedia.org/pipermail/qa/2014-May/001420.html

https://wiki.jenkins-ci.org/display/JENKINS/Claim+plugin

On each failling build page, a logged in user can 'claim' a build and
leave a comment explaining what is going on.  The comment can stick on
next failling builds until it is successful again.  That will ease
reviewing of browsertests failures.

Updates all browsertests jobs.

Bug: 64957
Change-Id: I9d09faeac4a9cd7a461fdaf8c97911b0443ace5d