Refactor doc publishing from labs
The tox 'doc' jobs are generating documentation in labs instance and
need to be synced to production doc.wikimedia.org.

This change use an intermediary instance in labs to rsync files (under
'docdest') to a determined path (based on ZUUL_UUID).  We can then
trigger a build on the 'doc' host gallium to fetch from the intermediary
rsync server and publish the files at doc.wikimedia.org/{docdest}

Meant for tox doc jobs for now, needs to be generalized somehow.

Creates a shared job that will run on the master. Only keeps 15 days of
build history:
    publish-doc

Updates:
    mw-tools-releng-tox-doc-publish
    mw-tools-scap-tox-doc-publish

Change-Id: Ia611474359eec618d8fbc1b551abc8afb504e798