defaults: Make values for 'wipe-workspace' and 'clean' explicit
This changes nothing but since we hardcode them everywhere else,
it makes it a lot easier to understand what is going on because
unlike all other jobs, mwext jobs are the only ones that use
'wipe' instead of 'clean'. And this wasn't obvious since it
was using that via the Jenkins-Git-Plugin default value.

Bug: 66429
Change-Id: I2d75c8bca250af663e3eddcf4d96cbd50024c768