Replace puppet-validate job with pplint-HEAD
The operations-puppet-validate job ends up parsing all manifests and use
a rakefile which is not very safe (though we have a workaround).
Additionally it could only run on gallium which is a slowdown.

Instead switch to the pplint-HEAD template that looks for .pp files
changed in HEAD and xargs them to puppet parser validate.

Change-Id: Ie3ab603f01b383b5224f056a3ec11bfd5f92781e