Trigger pywikibot-core-tox-nose
John Vandenberg	and others did a nice work at making it possible to run
a subset of pywikibot unit tests without having a configuration file nor
ever hitting a live site.

This move the job pywikibot-core-tox-nose to be triggered when
whitelisted people propose a patch and on gating.

Depends on:
https://gerrit.wikimedia.org/r/#/c/152959/

Change-Id: Iadf1bb6d6cc4230609b24e07a67309a6d3ccfe91