beta-scap-eqiad: rm attached build.log, use HTML
The job email notification attach the build.log to the email which is
not that useful and can cause the email to bounce when it ends up being
too large.  Recipients can just look at the console log instead.

Also have the email formatted as HTML.  Might end up being nicer to the
eyes.

Change-Id: I908c8a63acd61cb5bac99566cfda37bcabe77976