MEDIAWIKI_API_UPLOAD_URL is now a build parameter
The language screenshot job is hardcoded to points to the production
commons wiki.  We might want to manually trigger the job and have
the resulting screenshots uploaded on a different wiki (such as the
beta cluster commons wiki).

Make MEDIAWIKI_API_UPLOAD_URL a build parameter defaulting to
https://commons.wikimedia.org/w/api.php . Jenkins export it in the
environment.

This is for the job:
browsertests-VisualEditor-language-screenshot-linux-firefox-sauce

Change-Id: Idd630915b4631b12c33ab0bbcad4311dda77022c