Group MediaWiki jobs in the same gate queue
Repositories depending on mediawiki/core.git should all be in the same
dependent pipeline (the gate-and-submit with queue name 'mediawiki').

This way, whenever a patch for core enter the queue and can potentially
break a patch that has been enqueued later on, we will properly fails.

Use the dummy mediawiki-gate.

Bug: 58772
Change-Id: I3d4298695216016bea0ee437788e03597e54fe26