VisualEditor qunit job now process submodules
The VisualEditor qunit job mwext-VisualEditor-qunit ends up being broken
because it fetch the VisualEditor project has a submodule under /lib/ve.

Replace the default job template '{name}-{ext-name}-qunit' which does
not process submodules with a job template that does and has the same
name: 'mwext-VisualEditor-qunit'.

We do a `git clean -xqdf` to ensure the submodule has nothing left from
a previous checkout.  With Ubuntu Trusty, `git submodule update` learned
the --force option to handle that, but it is not available on the git
version of Ubuntu Precise :-/

Change-Id: I6f2253abf081ec47864f4744ae3df7ac38966584