Add macros for asserting node.js version
We now have slaves using Ubuntu Precise or Trusty, each providing a
different node minor version (respectively 0.8.x and 0.10.x).  We tie
jobs to them using the labels UbuntuPrecise and UbuntuTrusty which
should be fine until we make a configuration change.

To make sure we always run the job with the proper version, this patch
provide an assertion macro which would let one test the node version
being used and fails the build early if it does not meet
expectation.

The macro usage is straightforward:

    builders:
     - assert-node-version:
       version: ^0\.8\.

The `version` parameter is passed to `grep` which by default use basic
regular expressions.

Provide wrappers to easily assert node v0.8.x and v0.10.x

Change-Id: I5099616189dd7a26fae556d982b9669ff835923d