Normalize mediawiki/core phplint job
When linting mediawiki/core, we want to avoid submodules present in wmf
branches.  To do we went with a specific template '{name}-lint' which
overriden the default git scm to prevent submodule processing.

Nowadays, '{name}-phplint' does not process submodules either so switch
to it and get rid of the '{name}-lint' job template.

Summary of the change:

* job is renamed from mediawiki-core-lint to mediawiki-core-phplint
  (requires a chnage in Zuul config)
* build history is bumped from 15 days to 90 days
* Git remote URL and reference repo now uses $ZUUL_PROJECT instead of
  hardcoded mediawiki/core

Change-Id: I735b7ac59dd89315d61f1ea207ad0dca7c09a092