Remove VisualEditor-qunit
On review of JJB change https://gerrit.wikimedia.org/r/#/c/163837/ the
job should only be triggered for mediawiki/extensions/VisualEditor which
is where the integration work is being done.

It is also very confusing to have it named 'VisualEditor-qunit' which
would suggest it is meant to be triggered on upstream repository
VisualEditor/VisualEditor when it really tests the MediaWiki extension.

Change-Id: I6218a0bd3d06d13b794b5c03dbf4b4e064cc0193