Revert "Don't run oojs-ui-phplint for changes without php files"
This cause pipelines having a phplint as a parent job to be blocked:

Example for mediawiki/core test pipeline:

   - mediawiki-core-phplint:
        - mediawiki-vendor-integration
        - mediawiki-core-qunit

If a change is proposed without any php file change, the phplint
is not triggered and the child jobs are not either.

That cause the change to be locked in the queue.


This reverts commit 5e905e7c9dde9f47482de84e725ee48388c1dc43.

Change-Id: Id00c25ab7ad96fb9ff7c939a8d8e1a86651134a9