Use templates for all Wikidata jobs
Some jobs require tests with specific tags to be executed.
For this to work I needed to introduce another parameter to
specify custom tags. Called it cucumber_tags. If it is not set
everything behaves as before. If it is set, only tests with the
specified tags will be executed.

Change-Id: I7d7e528024c849b939cb5b00a2d522ec39d50ff1