Namespace GEM_HOME based on ruby version
A job migrated to Trusty/ruby2.0 ended up failling with a segfault in
the json gemfile. It uses some compiled code which is hold in a per
instance cache by setting GEM_HOME.   Since that cache is populated via
ruby 1.9.x, the compiled version is not going to work with ruby 2.x

Namespace the per instance cache (GEM_HOME) by using the ruby version.
That should be good enough.

Change-Id: Ibc06e59966c37dbfdda6f30a03f2e12b36eb60e4