Update patch set 11
Patch Set 11: Code-Review+1

Good enough for now.

I eventually want the bundle macro to support extra arguments to be passed and that we could set in the job-template.  But then such a parameter will always be required which is cumbersome.  I have a patch for JJB, will propose it upstream.


Meanwhile, this change looks fine to me.  The rubocop output is only listed in the console for now, but that is acceptable.   As stated on IRC, we can improve later on.

Patch-set: 11
Label: Code-Review=+1