Update patch set 1
Patch Set 1:

During my last vacations, I made integration-zuul-layoutvalidation fails whenever Zuul is configured to trigger jobs that are not deployed on Jenkins.

So whenever we propose a change introducing new jobs, it will fails indeed.

I guess that feature should be disabled on patchset submission but enabled on +2 to remember folks to deploy the jobs.


Workarounds:
- deploy the jobs and recheck :-D
- ignore the result, work on patch and once deployed +2 it and it will pass.

Patch-set: 1