Only test Jenkins jobs existence on gate-and-submit
Whenever someone propose a change that introduces new jobs, the
integration-zuul-layoutvalidation jobs fails because the job are not
found in Jenkins.  That is to be expected since the change is proposing
to add them.

The failure has confused a few people recently. I originally introduced
it to prevent Zuul configuration changes to be merged/deployed while
jobs are not deployed.

Remove the validation against Jenkins when patches are proposed (test
pipeline).
On gate, use a new job that does assert the jobs exists.

This way, if someone forget to deploy the jobs but +2 the change, it
will not land remembering people to do the deploy.

Change-Id: Id8f8eecc5ded1a77b120df090db9100f142b99d9